+++ trlevel/Level_tr3_psx.cpp
@@ -9,12 +9,6 @@ namespace trlevel
 {
     namespace
     {
-        template <typename T>
-        void write(std::basic_ospanstream<uint8_t>& file, const T& value)
-        {
-            file.write(reinterpret_cast<const uint8_t*>(&value), sizeof(value));
-        }
-
         uint32_t get_skybox_id(PlatformAndVersion version)
         {
             if (is_tr3_ects(version))
@@ -131,112 +125,6 @@ namespace trlevel
             read_room_reverb_info(activity, file, room);
             skip(file, 1); // filler in TR3
         }
-
-        std::vector<uint8_t> convert_vag_to_wav(const std::vector<uint8_t>& bytes)
-        {
-            std::span span{ reinterpret_cast<const uint8_t*>(&bytes[0]), bytes.size()};
-            std::basic_ispanstream<uint8_t> in_stream{ span };
-            in_stream.exceptions(std::istream::failbit | std::istream::badbit | std::istream::eofbit);
-
-            // Skip past the name
-            in_stream.seekg(16, std::ios::beg);
-
-            double s_1 = 0.0;
-            double s_2 = 0.0;
-            double samples[28];
-
-            double f[5][2] = { { 0.0, 0.0 },
-                    {   60.0 / 64.0,  0.0 },
-                    {  115.0 / 64.0, -52.0 / 64.0 },
-                    {   98.0 / 64.0, -55.0 / 64.0 },
-                    {  122.0 / 64.0, -60.0 / 64.0 } };
-
-            std::vector<uint8_t> out_bytes;
-
-            // , shift_factor, flags;
-            while (static_cast<uint32_t>(in_stream.tellg()) + 48 < static_cast<uint32_t>(bytes.size()))
-            {
-                if (static_cast<uint32_t>(in_stream.tellg()) == bytes.size())
-                {
-                    break;
-                }
-
-                int predict_nr = read<char>(in_stream);
-                int shift_factor = predict_nr & 0xf;
-                predict_nr >>= 4;
-                uint8_t flags = read<uint8_t>(in_stream);
-                if (flags == 7)
-                {
-                    break;
-                }
-
-                for (int i = 0; i < 28; i += 2)
-                {
-                    int d = read<uint8_t>(in_stream);
-
-                    int s = (d & 0xf) << 12;
-                    if (s & 0x8000)
-                    {
-                        s |= 0xffff0000;
-                    }
-                    samples[i] = static_cast<double>(s >> shift_factor);
-
-                    s = (d & 0xf0) << 8;
-                    if (s & 0x8000)
-                    {
-                        s |= 0xffff0000;
-                    }
-                    samples[i + 1] = static_cast<double>(s >> shift_factor);
-                }
-
-                for (int i = 0; i < 28; i++) {
-                    samples[i] = samples[i] + s_1 * f[predict_nr][0] + s_2 * f[predict_nr][1];
-                    s_2 = s_1;
-                    s_1 = samples[i];
-
-                    int d = static_cast<int>(samples[i] + 0.5);
-                    
-                    // fputc(d & 0xff, pcm);
-                    // fputc(d >> 8, pcm);
-                    out_bytes.push_back(static_cast<uint8_t>(d & 0xff));
-                    out_bytes.push_back(static_cast<uint8_t>(d >> 8));
-                }
-            }
-
-            uint32_t sample_frequency = 11025;
-
-            std::vector<uint8_t> results;
-            results.resize(500000);
-
-            std::span out_span{ reinterpret_cast<uint8_t*>(&results[0]), results.size() };
-            std::basic_ospanstream<uint8_t> out_stream{ out_span };
-            out_stream.exceptions(std::istream::failbit | std::istream::badbit | std::istream::eofbit);
-
-            out_stream << "RIFF";
-            out_stream.seekp(4, std::ios::cur);
-            out_stream << "WAVEfmt ";
-            write(out_stream, 16);
-            write<uint16_t>(out_stream, 1);
-            write<uint16_t>(out_stream, 1);
-            write<uint32_t>(out_stream, sample_frequency);
-            write<uint32_t>(out_stream, sample_frequency * 2);
-            write<uint16_t>(out_stream, 2);
-            write<uint16_t>(out_stream, 16);
-
-            out_stream << "data";
-            out_stream.seekp(4, std::ios::cur);
-
-            out_stream.write(&out_bytes[0], out_bytes.size());
-
-            uint32_t file_size = static_cast<uint32_t>(out_stream.tellp());
-            out_stream.seekp(4, std::ios::beg);
-            write<uint32_t>(out_stream, file_size - 8);
-            out_stream.seekp(40, std::ios::beg);
-            write<uint32_t>(out_stream, file_size - 44);
-
-            results.resize(file_size);
-            return results;
-        }
     }
 
     void Level::generate_mesh_tr3_psx(tr_mesh& mesh, std::basic_ispanstream<uint8_t>& stream)
@@ -344,33 +232,8 @@ namespace trlevel
 
     void Level::load_tr3_psx(std::basic_ispanstream<uint8_t>& file, trview::Activity& activity, const LoadCallbacks& callbacks)
     {
-        auto sound_offsets = read_vector<uint32_t, uint32_t>(file);
-        auto sound_data = read_vector<uint32_t, byte>(file);
-
-        for (uint32_t s = 0; s < sound_offsets.size(); ++s)
-        {
-            const std::size_t offset = sound_offsets[s];
-            const std::size_t size = s == sound_offsets.size() - 1 ? 
-                sound_data.size() - offset - 1:
-                sound_offsets[s + 1] - offset;
-            size;
-            const auto converted = convert_vag_to_wav({ &sound_data[offset], &sound_data[offset + size] });
-
-            // TODO: Remove temporary dump
-            std::ofstream out_file;
-            out_file.open(std::format("C:\\dev\\trview-temp\\sounds\\{}.wav", s), std::ios::binary | std::ios::out);
-            out_file.write(reinterpret_cast<const char*>(&converted[0]), converted.size());
-            out_file.close();
-
-            callbacks.on_sound(static_cast<uint16_t>(s), converted);
-        }
-
+        read_sounds_psx(activity, file, callbacks, 11025);
         const bool ects = is_tr3_ects(_platform_and_version);
-
-        // TR3:
-        // Num samples(uint32) at offset 0x04, followed by an array of raw sample offsets(uint32[num samples]) and the combined sample size(in bytes).
-        // Raw sample data then follows consecutively.
-        // Note: This format also applies to the TR3 ECTS demo and later beta builds.
 
         for (int i = 0; i < (ects ? 10 : 13); ++i)
         {
